requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
numpy==1.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.6
websocket-client==1.1.0
websockets==9.1
whitenoise==5.3.0
zope.event==4.5.0
zope.interface==5.4.0
